разбивает на модули (#8)
