+++ js/photo-data.js
@@ -1,0 +1,75 @@
+const POST_QUANTITY = 25;
+
+const COMMENTS_QUANTITY = {
+  min: 0,
+  max: 30
+};
+
+const PHOTO_DESCRIPTIONS = [
+  'Как круто!',
+  'Обожаю это фото',
+  'Очень красиво смотрится',
+  'Настоящее вдохновение',
+  'Сразу чувствуется твоя энергия',
+  'Просто невероятно',
+  'Очень атмосферное фото',
+  'Ты как всегда на высоте',
+  'Настроение передается на все сто',
+  'Какая классная атмосфера',
+  'Так держать',
+  'Завораживает',
+  'Вот это уровень',
+  'Очень круто получилось',
+  'Видно, что от души',
+  'Очень стильно',
+  'Тебе это идет',
+  'Гармония и красота',
+  'Настоящий эстет',
+  'Фантастическое фото',
+  'Такой легкий кадр',
+  'Прямо в самое сердце',
+  'Словно кадр из фильма',
+  'Как будто из мечты',
+  'Очень вдохновляющий кадр',
+];
+
+const LIKES_QUANTITY = {
+  min: 15,
+  max: 200,
+};
+
+const AVATAR_QUANTITY = {
+  min: 1,
+  max: 6,
+};
+
+const COMMENT_MESSAGES = [
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const USER_NAMES = [
+  'Иван',
+  'Анна',
+  'Мария',
+  'Алексей',
+  'Екатерина',
+  'Сергей',
+  'Ольга',
+  'Михаил',
+  'Виктория',
+  'Дмитрий',
+];
+
+export {
+  POST_QUANTITY,
+  COMMENTS_QUANTITY,
+  PHOTO_DESCRIPTIONS,
+  LIKES_QUANTITY,
+  AVATAR_QUANTITY,
+  COMMENT_MESSAGES,
+  USER_NAMES
+};
